+++ mobile/composeApp/src/commonMain/kotlin/io/asterixorobelix/afrikaburn/models/Artist.kt
@@ -1,0 +1,9 @@
+package io.asterixorobelix.afrikaburn.models
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class Artist(
+    @SerialName("s") val name: String = ""
+)
