+++ mobile/composeApp/src/commonMain/kotlin/io/asterixorobelix/afrikaburn/models/TabDataSource.kt
@@ -1,0 +1,6 @@
+package io.asterixorobelix.afrikaburn.models
+
+data class TabDataSource(
+    val fileName: String,
+    val displayName: String
+)
